medium/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-jsonapi==7.0.2
inflection==0.5.1
python-dotenv==1.0.1
sqlparse==0.5.1
asgiref==3.8.1
Django==5.1.1
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-jsonapi==7.0.2
djangorestframework-simplejwt==5.3.1
inflection==0.5.1
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
